changed some variables to size_t and so on

+++ includes/graph.h
@@ -6,7 +6,7 @@
 /*   By: npimenof <npimenof@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/29 11:59:05 by npimenof          #+#    #+#             */
-/*   Updated: 2020/10/06 11:35:52 by npimenof         ###   ########.fr       */
+/*   Updated: 2020/10/06 15:16:16 by npimenof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct		s_adjlist
 typedef struct		s_graph
 {
 	t_adjlist		adj;
-	int				vertices;
+	size_t			vertices;
 }					t_graph;
 
 typedef struct		s_node
@@ -44,7 +44,7 @@ typedef struct		s_edge
 	struct s_edge	*residual;
 }					t_edge;
 
-t_graph				*init_graph(int vertices);
+t_graph				*init_graph(size_t vertices);
 t_adjlist			*to_adj(t_graph *g);
 size_t				get_node_index(t_node *n);
 int					add_edge(t_graph *g, t_node *s, t_node *d);
